+++ src/cli.ts
@@ -1,29 +1,12 @@
 import { readFileSync } from 'fs';
 import parse from '.';
+import { formatChangelog } from './formatter';
+import plainTextFormatter from './formatters/plain-text';
 
 const filename = process.argv[2];
 
 const contents = readFileSync(filename);
 parse(contents.toString())
   .then((parsed) => {
-    parsed.releases.forEach((release) => {
-      // todo: extract this into a util to format releases
-      process.stdout.write(release.version || 'unreleased');
-      process.stdout.write(' (');
-      process.stdout.write(release.date || 'no date');
-      process.stdout.write(')\n');
-
-      function printItems(prefix: string, items: string[] | undefined) {
-        items?.forEach((item: string) => {
-          process.stdout.write(`- ${prefix}: ${item}\n`);
-        });
-      }
-
-      printItems('Added', release.added);
-      printItems('Changed', release.changed);
-      printItems('Deprecated', release.deprecated);
-      printItems('Removed', release.removed);
-      printItems('Fixed', release.fixed);
-      printItems('Security', release.security);
-    });
+    process.stdout.write(formatChangelog(parsed, plainTextFormatter()));
   });